+++ Modules/booking/Model/BkCalSupInfo.php
@@ -148,7 +148,6 @@ class BkCalSupInfo extends Model {
      */
     public function delete($id_space, $id) {
         $sql = "UPDATE bk_calsupinfo SET deleted=1,deleted_at=NOW() WHERE id=? AND id_space=?";
-        // $sql = "DELETE FROM bk_calsupinfo WHERE id = ? AND id_space=?";
         $this->runRequest($sql, array($id, $id_space));
     }