+++ Modules/invoices/Model/InVisa.php
@@ -112,7 +112,6 @@ class InVisa extends Model {
      */
     public function delete($id_space, $id) {
         $sql = "UPDATE in_visa SET deleted=1,deleted_at=NOW() WHERE id=? AND id_space=?";
-        // $sql = "DELETE FROM in_visa WHERE id = ? AND id_space=?";
         $this->runRequest($sql, array($id, $id_space));
     }