+++ Modules/resources/Model/ReEvent.php
@@ -97,7 +97,6 @@ class ReEvent extends Model {
      */
     public function delete($id_space, $id) {
         $sql = "UPDATE re_event SET deleted=1,deleted_at=NOW() WHERE id=? AND id_space=?";
-        // $sql = "DELETE FROM re_event WHERE id = ? AND id_space=? AND deleted=0";
         $this->runRequest($sql, array($id, $id_space));
     }