+++ Modules/booking/Model/BkBookingSettings.php
@@ -11,6 +11,10 @@ require_once 'Modules/core/Model/CoreSpace.php';
  */
 class BkBookingSettings extends Model {
 
+    public function __construct() {
+        $this->tableName = "bk_booking_settings";
+    }
+
     /**
      * Create the booking settings entry table
      *
@@ -214,7 +218,7 @@ class BkBookingSettings extends Model {
             } elseif ($entryList[$i]['font'] == "italic") {
                 $summary .= "</em>";
             }
-            if ($last == false) {
+            if ($summary && $last == false) {
                 if ($displayHorizontal) {
                     $summary .= " ";
                 } else {