+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## 2.4.1 [unreleased]
 
+* [core] add missing tableName properties in model constructors
+* [space] improve todolist sql resquests
 * [helpdesk] close connection in anycase
 * [core] fix TODO in corespaceadmin, do not get all bookings, just count them
 * [core] if module is inactive, fix authorization (isUserMenuSpaceAuthorized)