+++ CHANGELOG.md
@@ -2,11 +2,15 @@
 
 ## 2.4.1 [unreleased]
 
+* [core] add missing tableName properties in model constructors
+* [space] improve todolist sql resquests
 * [helpdesk] close connection in anycase
 * [core] fix TODO in corespaceadmin, do not get all bookings, just count them
 * [core] if module is inactive, fix authorization (isUserMenuSpaceAuthorized)
 * [core] disable timezone, impacts existing bookings recorded as UTC
 * [booking] fix calendar when area has no resource
+* [booking] fix calendar display when resa spans multiple days
+* [booking] fix calendar month display for month overlap resa
 
 ## 2.4.0